+++ torch/nn/parallel/replicated_tensor_ddp_utils.py
@@ -1,0 +1,21 @@
+from contextlib import contextmanager
+
+_DDP_WITH_REPLICATED_TENSOR = False
+
+@contextmanager
+def _ddp_replicated_tensor(val):
+    global _DDP_WITH_REPLICATED_TENSOR
+    old_val = _DDP_WITH_REPLICATED_TENSOR
+    _DDP_WITH_REPLICATED_TENSOR = val
+    try:
+        yield
+    finally:
+        _DDP_WITH_REPLICATED_TENSOR = old_val
+
+def _ddp_with_replicated_tensor_enabled():
+    global _DDP_WITH_REPLICATED_TENSOR
+    return _DDP_WITH_REPLICATED_TENSOR
+
+def _set_ddp_with_replicated_tensor(value):
+    global _DDP_WITH_REPLICATED_TENSOR
+    _DDP_WITH_REPLICATED_TENSOR = value
